Refactor find item logic into hook
A bit of code I came across that might not need to be at the component level
and can be pulled out into a Hook. Cleans up file and allows logic to be
reused elsewhere.
